fix: rename some files
