+++ apps/api/src/app/equipment/default-equipments.ts
